logic for 100 mile radius city query
this is the logic for the 100 mile radius city query.  I converted all
of the lat lon coords to radians and used a formula I derived from this
wikipedia article http://en.wikipedia.org/wiki/Great-circle_distance.
It assumes the earth is a sphere and that the radius is 4k miles, so
it’s crazy accurate, but I think it works for this instance.  The logic
in this controller is starting to get a little too much logic in it.  I
might move this to the model.