controller for new city user
this is the controller action for a new user.  I’m skipping the rails
standard verify authenticity token so that api requests can be made
from anywhere i.e. postman.  this method takes a user id and city and
adds it to the city user table.  no error handling yet